Added a playground for showing example usage
In the interest of not duplicating code, I put a symlink to
IntuitiveMath.swift in the playground’s Sources directory rather than
copying the file. The symlink uses the relative path to the file, so
the playground should work as long as it is in the same directory as
IntuitiveMath.swift.