Improved formatting of enum documentation
This is necessary to get quick help to show documentation for the
individual enum cases. As far as I can tell there’s no way to get both
documentation on individual cases and documentation on every case in
the whole enum’s quick help without duplicating documentation, and
given that decision this option seems preferable to me. (It also
appears to be consistent with how Apple documents enums in Swift.)