Deleted the playground for now
I haven’t added anything to it yet and including IntuitiveMath.swift in
the playground’s Sources as a symbolic link isn’t working particularly
well. This SO answer mentions that hard links may work better for this:
http://stackoverflow.com/a/24088624/2680007 …but I realize I may want
to turn this into an actual module/framework instead (as per Rick’s
answer on the same SO thread). I’ll continue developing this in the
project I’m using it in for now, pushing changes here, and perhaps down
the road turn it into a its own little library.